create toJsonSerializationDefaultOptions
